+++ pandas_gbq/gbq.py
@@ -650,6 +650,9 @@ def _bqschema_to_nullsafe_dtypes(schema_fields):
     # See:
     # http://pandas.pydata.org/pandas-docs/dev/missing_data.html
     # #missing-data-casting-rules-and-indexing
+    #
+    # If you update this mapping, also update the table at
+    # `docs/source/reading.rst`.
     dtype_map = {
         "FLOAT": np.dtype(float),
         # Even though TIMESTAMPs are timezone-aware in BigQuery, pandas doesn't